backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
langchain
langchain-community
faiss-cpu
sentence-transformers
celery
redis
httpx
python-dotenv
pypdf
paddlepaddle
paddleocr
pillow
opencv-python-headless
numpy
transformers
torch
accelerate
